# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MEX API",
    "version": "1.0.0",
    "description": "API documentation of medical examination system (T5)"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/api/mex/{patient_id}": {
      "get": {
        "tags": ["Medical examination record"],
        "summary": "Retrieve all medical examination (MEX) records for a specific patient",
        "description": "Fetches a list of MEX records for a given patient ID (hn), including examination details and related prescriptions.",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The hospital number (hn) of the patient to fetch MEX records."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the MEX records.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "hn": {
                        "type": "integer",
                        "example": 123456
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24T08:00:00Z"
                      },
                      "dx": {
                        "type": "string",
                        "example": "Viral infection"
                      },
                      "RxList": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "medicalName": {
                              "type": "string",
                              "example": "Paracetamol"
                            },
                            "quantity": {
                              "type": "integer",
                              "example": 10
                            },
                            "note": {
                              "type": "string",
                              "example": "might make your drowsy"
                            },
                            "unit": {
                              "type": "string",
                              "example": "tablets"
                            },
                            "mexId": {
                              "type": "integer",
                              "example": 1
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to fetch MEX records due to an error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error na kub"
                    },
                    "error": {
                      "type": "object",
                      "description": "Error details."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/mex/{id}": {
      "put": {
        "tags": ["Medical examination record"],
        "summary": "Update a medical examination (MEX) record by ID",
        "description": "Updates an existing MEX record with new data for a specific ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the MEX record to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "docID": {
                    "type": "integer",
                    "example": 123
                  },
                  "hn": {
                    "type": "string",
                    "example": "HN123456"
                  },
                  "chiefComplaint": {
                    "type": "string",
                    "example": "Headache"
                  },
                  "presentHistory": {
                    "type": "string",
                    "example": "Patient has a history of fever."
                  },
                  "dx": {
                    "type": "string",
                    "example": "Viral infection"
                  },
                  "procedure": {
                    "type": "string",
                    "example": "Physical examination"
                  },
                  "appointment": {
                    "type": "string",
                    "example": "2025-02-26T10:00:00Z"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-02-24T08:00:00Z"
                  },
                  "examination_filename": {
                    "type": "string",
                    "example": "Johny-examfile.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the MEX record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updatedMEX": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "docID": {
                          "type": "integer",
                          "example": 123
                        },
                        "hn": {
                          "type": "string",
                          "example": "HN123456"
                        },
                        "chiefComplaint": {
                          "type": "string",
                          "example": "Headache"
                        },
                        "presentHistory": {
                          "type": "string",
                          "example": "Patient has a history of fever."
                        },
                        "dx": {
                          "type": "string",
                          "example": "Viral infection"
                        },
                        "procedure": {
                          "type": "string",
                          "example": "Physical examination"
                        },
                        "appointment": {
                          "type": "string",
                          "example": "2025-02-26T10:00:00Z"
                        },
                        "date": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-02-24T08:00:00Z"
                        },
                        "examination_filename": {
                          "type": "string",
                          "example": "Johny-examfile.pdf"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to update MEX record due to an error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error updating MEX"
                    },
                    "error": {
                      "type": "object",
                      "description": "Error details."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}